+++ lib/flutter_device_ring.dart
@@ -1,0 +1,2 @@
+export 'src/device_ring.dart';
+export 'src/device_ring_theme.dart';